# small refactory on cache memory